azureml_upload/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.27
python-dotenv==1.0.0
lxml==5.1.0
fake-useragent==1.4.0
selenium==4.18.1
webdriver-manager==4.0.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.20.0 